# src/aegis_range/openapi/aegis-bank.json
{
  "openapi": "3.1.0",
  "info": {"title": "Aegis Bank API", "version": "1.0.0"},
  "servers": [{"url": "http://aegis-bank:8101"}],
  "paths": {
...
    "/api/accounts/{account_id}": {"get": {"operationId": "getAccount", "summary": "Get account details", "security": [{"bearerAuth": []}], "parameters": [{"name": "account_id", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "Account"}, "403": {"description": "Access denied"}, "404": {"description": "Account not found"}}}},
...
  },
  "components": {"securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer"}}}
}